+++ src/Tests/Behavioral/ErrorfFormatting/main.cs
@@ -6,8 +6,8 @@ partial class main_package {
 
 internal static void Main() {
     fmt.Println(fmt.Errorf("plain message"u8));
-    fmt.Println(fmt.Errorf("got %v"u8, 42));
-    fmt.Println(fmt.Errorf("name %s = %d"u8, "x", 7));
+    fmt.Println(fmt.Errorf("got %v"u8, (nint)(42)));
+    fmt.Println(fmt.Errorf("name %s = %d"u8, "x", (nint)(7)));
     fmt.Println(fmt.Errorf("%v and %v"u8, true, "y"));
     ((Action)(() => func((defer, recover) => {
         defer(() => {
